add aws rdb configuration in application-dev.properties and change docker-compose.yml

+++ src/main/java/com/example/demo/student/StudentService.java
@@ -1,5 +1,7 @@
 package com.example.demo.student;
 
+import com.example.demo.student.exception.BadRequestException;
+import com.example.demo.student.exception.StudentNotFoundException;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -16,10 +18,22 @@ public class StudentService {
 
     public void addStudent(Student student) {
         // Check if email is valid or not
+        Boolean existsEmail = studentRepository.selectExistsEmail(student.getEmail());
+        if (existsEmail) {
+            throw new BadRequestException(
+                    "Email" + student.getEmail() + " taken"
+            );
+        }
+
         studentRepository.save(student);
     }
 
     public void deleteStudent(Long studentId) {
+        if (!studentRepository.existsById(studentId)) {
+            throw new StudentNotFoundException(
+                    "Student with id" + studentId + " does not exists"
+            );
+        }
         studentRepository.deleteById(studentId);
     }
 }
